removing Fabric::Gateway namespace in favor of just Fabric; adding unit testing and CI
