+++ .github/workflows/rspec.yml
@@ -33,3 +33,5 @@ jobs:
         bundler-cache: true # runs 'bundle install' and caches installed gems automatically
     - name: Run tests
       run: bundle exec rake
+      env:
+        CODECOV_TOKEN: 4b75ccca-3149-4610-99fb-3a5aeff4ff00
