+++ lib/fabric/constants.rb
@@ -1,0 +1,6 @@
+module Fabric
+  module Constants
+    ## Variables
+    CHANNEL_HEADER_VERSION = 1
+  end
+end
