--- lib/fabric/gateway/constants.rb
@@ -1,20 +1,0 @@
-module Fabric
-  module Gateway
-    module Constants
-      ## System Chaincodes
-      LSCC = 'lscc'.freeze
-      QSCC = 'qscc'.freeze
-      CSCC = 'cscc'.freeze
-
-      ## System Channels
-      SYSTEM_CHANNEL_NAME = 'testchainid'.freeze
-
-      ## System Functions
-      FUNC_GET_CHANNELS = 'GetChannels'.freeze
-      FUNC_GET_CONFIG_BLOCK = 'GetConfigBlock'.freeze
-
-      ## Variables
-      CHANNEL_HEADER_VERSION = 1
-    end
-  end
-end
