--- lib/fabric/gateway/version.rb
@@ -1,5 +1,0 @@
-module Fabric
-  module Gateway
-    VERSION = "0.1.0"
-  end
-end
