+++ lib/fabric/network.rb
@@ -1,0 +1,3 @@
+module Fabric
+  
+end
