+++ lib/fabric/version.rb
@@ -1,0 +1,3 @@
+module Fabric
+  VERSION = "0.1.0"
+end
