--- spec/fabric/gateway_spec.rb
@@ -1,9 +1,0 @@
-RSpec.describe Fabric::Gateway do
-  it "has a version number" do
-    expect(Fabric::Gateway::VERSION).not_to be nil
-  end
-
-  it "does something useful" do
-    expect(false).to eq(true)
-  end
-end
